Refactor handles step 4: move store/get + app_settings into data_stores
- Add a `settings_store` to data_stores: a `SettingsStore` port + `JsonSettingsStore`
  adapter (JSON file), moved from `handles::app_settings`.
- Inline the icon read+resize (formerly `handles::store::get::resource_icons`) into
  `ResourceData`: read raw bytes from the IconsDb store, resize via the icon provider.
- Drop the dead `accounts_and_resources` helper; delete `handles::store` and
  `handles::app_settings`.
