+++ 01_ports_and_adapters/data_stores/src/adapters/mod.rs
@@ -1,3 +1,6 @@
 mod data_store;
 mod icon_data_store;
+mod settings_store;
 mod wallet_data_store;
+
+pub use settings_store::JsonSettingsStore;
