+++ 01_ports_and_adapters/data_stores/src/lib.rs
@@ -1,2 +1,5 @@
 pub mod adapters;
 pub mod ports;
+
+pub use adapters::JsonSettingsStore;
+pub use ports::settings_store::SettingsStore;
