+++ 01_ports_and_adapters/data_stores/src/ports/mod.rs
@@ -1,3 +1,4 @@
 pub mod data_store;
 pub mod icon_data_store;
+pub mod settings_store;
 pub mod wallet_data_store;
