+++ 3_handles/src/lib.rs
@@ -1,5 +1,3 @@
-pub mod app_settings;
 pub mod database_handle;
 pub mod statics;
-pub mod store;
 pub mod wallet;
